remove capabilities

+++ src/main/java/net/maplecraft/effect/BuffSpeedBoostMobEffect.java
@@ -8,7 +8,7 @@ import net.minecraft.world.entity.ai.attributes.Attributes;
 public class BuffSpeedBoostMobEffect extends MapleMobEffect {
     public BuffSpeedBoostMobEffect() {
         super(MobEffectCategory.BENEFICIAL, 0xff8103);
-        this.addAttributeModifier(Attributes.MOVEMENT_SPEED, "11AEAC56-376B-4498-9A5B-123468517682", (double)0.001D, AttributeModifier.Operation.ADDITION);
+        this.addAttributeModifier(Attributes.MOVEMENT_SPEED, "11AEAC56-376B-4498-9A5B-123468517682", 0.001D, AttributeModifier.Operation.ADDITION);
     }
 
     @Override
