+++ src/main/java/net/maplecraft/effect/EquipAttackPercentBoostMobEffect.java
@@ -9,7 +9,7 @@ public class EquipAttackPercentBoostMobEffect extends MapleMobEffect {
     public static int equipAttackPercentBoostBase = 5;
     public EquipAttackPercentBoostMobEffect() {
         super(MobEffectCategory.BENEFICIAL, 0xff8103);
-        this.addAttributeModifier(Attributes.ATTACK_DAMAGE, "325D7064-6A63-1F59-CABE-C2C23A517682", (double)0.05D, AttributeModifier.Operation.MULTIPLY_BASE);
+        this.addAttributeModifier(Attributes.ATTACK_DAMAGE, "325D7064-6A63-1F59-CABE-C2C23A517682", 0.05D, AttributeModifier.Operation.MULTIPLY_BASE);
     }
 
     @Override
