+++ src/main/java/net/maplecraft/effect/EquipHealthBoostMobEffect.java
@@ -8,7 +8,7 @@ import net.minecraft.world.entity.ai.attributes.Attributes;
 public class EquipHealthBoostMobEffect extends MapleMobEffect {
     public EquipHealthBoostMobEffect() {
         super(MobEffectCategory.BENEFICIAL, 0xff8103);
-        this.addAttributeModifier(Attributes.MAX_HEALTH, "5D6F0BA2-1186-46AC-B896-C61C5C517682", (double)1.0D, AttributeModifier.Operation.ADDITION);
+        this.addAttributeModifier(Attributes.MAX_HEALTH, "5D6F0BA2-1186-46AC-B896-C61C5C517682", 1.0D, AttributeModifier.Operation.ADDITION);
     }
 
     @Override
